changed hist_mean from float to double
Because mean() returns a double.

+++ ExtractFeat.h
@@ -13,7 +13,7 @@ using namespace std;
 
 struct Fillet {
 	String name;
-	float hist_mean[2] = { 0 };						// BGR
+	double hist_mean[2] = { 0 };						// BGR
 	double area = 0, rectangularity=0, hullarity;						// Contour area + convexity(squarity) which is contour area divided by boundrect area.
 	float width = 0, height = 0;					// of RotatedRect
 	float skinArea = 0;
@@ -35,7 +35,7 @@ public:
 	void clearFileContent();
 	void displayImg(const String &name, const Mat &img);
 	void makeBinary(const Mat &img, Mat &bin);
-	//------------Nuværende-fisk----------------------
+	//------------NuvÃ¦rende-fisk----------------------
 	void getMeanHist(Fillet &fillet);
 	void getDimensions(Fillet &fillet);
 	void getBloodStains(Fillet &fillet);
